refactor FileAccess to always use kotlinx.io when possible
