+++ savedstate/src/jvmMain/kotlin/pro/respawn/flowmvi/savedstate/platform/FileAccess.jvm.kt
@@ -1,6 +1,7 @@
 package pro.respawn.flowmvi.savedstate.platform
 
 import kotlinx.coroutines.runInterruptible
+import kotlinx.io.files.SystemFileSystem
 import java.io.File
 import java.io.InputStream
 import java.util.zip.GZIPInputStream
@@ -31,17 +32,6 @@ internal actual suspend fun readCompressed(path: String): String? {
     return runInterruptible { file.inputStream().let(::GZIPInputStream).readOrNull() }
 }
 
-internal actual suspend fun write(data: String?, path: String) {
-    val file = File(path)
-    if (data == null) {
-        file.delete()
-        return
-    }
-    return runInterruptible { file.outputStreamOrEmpty().bufferedWriter().use { it.write(data) } }
-}
+internal actual suspend fun write(data: String?, path: String) = write(data, path, SystemFileSystem)
 
-internal actual suspend fun read(path: String): String? {
-    val file = File(path)
-    if (!file.exists()) return null
-    return runInterruptible { file.inputStream().readOrNull() }
-}
+internal actual suspend fun read(path: String): String? = read(path, SystemFileSystem)
